Refactor deactivate account ui page to several components
