+++ src/containers/deactivate-account-page/DeactivateAccountPage.module.css
@@ -42,27 +42,8 @@
 .button > span {
   font-size: var(--font-m) !important;
 }
-/** Confirm modal **/
-.container {
-  padding: 0 50px;
-}
-.container > * {
-  margin: 24px;
-}
-.confirmText {
-  text-align: center;
-  color: var(--accent-red);
-  font-weight: var(--font-demi-bold);
-}
-.spinner {
-  width: 48px;
-  height: 48px;
-  margin-left: auto;
-  margin-right: auto;
-}
 
 /** MOBILE **/
-
 .mobile.tile {
   padding: 24px;
   margin: 32px 8px;
@@ -76,22 +57,4 @@
 .mobile .button {
   margin: 12px 0 0;
   width: 100%;
-}
-
-/** Custom Drawer Title **/
-.drawerTitle {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding: 8px 24px 24px;
-}
-.drawerTitleHeader {
-  margin-bottom: 16px;
-  font-size: var(--font-l);
-  font-weight: var(--font-bold);
-}
-.drawerTitleWarning {
-  color: var(--accent-red);
-  font-weight: var(--font-demi-bold);
 }
