+++ src/containers/deactivate-account-page/components/desktop/DeactivateAccountConfirmationModal.module.css
@@ -1,0 +1,35 @@
+/** Confirm modal **/
+.container {
+  padding: 0 50px;
+}
+.container > * {
+  margin: 24px;
+}
+.confirmText {
+  text-align: center;
+  color: var(--accent-red);
+  font-weight: var(--font-demi-bold);
+}
+.spinner {
+  width: 48px;
+  height: 48px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.button {
+  height: auto !important;
+  margin-inline-end: 24px;
+  padding: 12px 24px !important;
+}
+.button.buttonDanger,
+.button.buttonDanger:hover {
+  background-color: var(--accent-red);
+}
+.button.buttonDanger:active {
+  background-color: var(--accent-red-dark-1);
+}
+
+.button > span {
+  font-size: var(--font-m) !important;
+}
