+++ src/containers/deactivate-account-page/components/desktop/DeactivateAccountPage.tsx
@@ -1,0 +1,35 @@
+import React from 'react'
+
+import Page from 'components/general/Page'
+import Header from 'components/general/header/desktop/Header'
+
+import {
+  messages,
+  DeactivateAccountPageProps
+} from '../../DeactivateAccountPage'
+
+import { DeactivateAccountConfirmationModal } from './DeactivateAccountConfirmationModal'
+
+export const DeactivateAccountPageDesktop = ({
+  children,
+  isConfirmationVisible,
+  isLoading,
+  onConfirm,
+  closeConfirmation
+}: DeactivateAccountPageProps) => {
+  return (
+    <Page
+      title={messages.title}
+      description={messages.description}
+      header={<Header primary={messages.title} />}
+    >
+      {children}
+      <DeactivateAccountConfirmationModal
+        isVisible={isConfirmationVisible}
+        onClose={closeConfirmation}
+        onConfirm={onConfirm}
+        isLoading={isLoading}
+      />
+    </Page>
+  )
+}
